// packages/config/src/roles.enum.ts
export enum Role {
  SUPER_ADMIN = 'SUPER_ADMIN',
  SUPPORT_ADMIN = 'SUPPORT_ADMIN',
  SYSTEM_AUDITOR = 'SYSTEM_AUDITOR',
  SCHOOL_OWNER = 'SCHOOL_OWNER',
  HEADMASTER = 'HEADMASTER',
  VICE_PRINCIPAL = 'VICE_PRINCIPAL',
  SCHOOL_ADMIN = 'SCHOOL_ADMIN',
  ACADEMIC_COORDINATOR = 'ACADEMIC_COORDINATOR',
  ADMIN_OFFICER = 'ADMIN_OFFICER',
  TEACHER = 'TEACHER',
  HEAD_OF_DEPARTMENT = 'HEAD_OF_DEPARTMENT',
  CLASS_TEACHER = 'CLASS_TEACHER',
  EXAM_OFFICER = 'EXAM_OFFICER',
  LAB_INSTRUCTOR = 'LAB_INSTRUCTOR',
  SUBSTITUTE_TEACHER = 'SUBSTITUTE_TEACHER',
  STUDENT = 'STUDENT',
  PARENT = 'PARENT',
  ACCOUNTANT = 'ACCOUNTANT',
  RECEPTIONIST = 'RECEPTIONIST',
  LIBRARIAN = 'LIBRARIAN',
  IT_ADMIN = 'IT_ADMIN',
}